refactor: create a reusable card component
